+++ pkg/beholder/durable_emitter.go
@@ -109,8 +109,9 @@ func DefaultDurableEmitterConfig() DurableEmitterConfig {
 // ingress outages.
 //
 // store is the persistence layer. Callers in a Postgres environment should pass
-// pgstore.New(ds); the indirection keeps the lib/pq driver out of consumers that
-// only need the beholder API (e.g. wasip1 builds of the workflow runtime).
+// durable_events.New(ds); the indirection keeps the lib/pq driver out of
+// consumers that only need the beholder API (e.g. wasip1 builds of the workflow
+// runtime).
 func SetupDurableEmitter(ctx context.Context, client *Client, store DurableEventStore, retransmit bool, lggr logger.Logger) error {
 	if client == nil {
 		return errors.New("beholder client not initialized")
